move common functionality to utils
